+++ PapersProject/src/Terms.scala
@@ -73,7 +73,8 @@ case class Reference(authors: List[Author], title: Title) extends Term {
   override def toString : String = authors.mkString("\n") + "\n--\n" + title
 }
 
-case class Link(index : Int, weight : Int) extends Term {
+//changed weight from Int to Double
+case class Link(index : Int, weight : Double) extends Term {
   override def toString : String = index + " " + weight
 }
